Updated the HTTP Post Edit action within the Student Controller.
Renamed the method given the signature conflict with another method and added an Attribute to indicate that it should still handle the Edit action.

Also adjusted the updating process so that it depends more on the actual Model as opposed to an arbitrary set of strings that supposedly match the Model.